# cmd/alpacascloud/web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Alpacas.cloud — Alpacas are awesome!",
    "description": "Share your love of alpacas! A [Signal](https://signal.org) bot is also available, just ask for an alpaca at this number: [+33771848935](tel:+33771848935).",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
    "/alpaca": {
      "get": {
        "summary": "Get a random alpaca",
        "operationId": "getAlpaca",
        "tags": [
          "alpaca"
        ],
        "parameters": [
          {
            "name": "width",
            "in": "query",
            "description": "Max width of the image",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          },
          {
            "name": "height",
            "in": "query",
            "description": "Max height of the image",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Could not find an alpaca matching your requirements"
          }
        }
      }
    },
    "/placeholder/{width}x{height}": {
      "get": {
        "summary": "Get a random alpaca placeholder",
        "operationId": "getPlaceholder",
        "tags": [
          "alpaca"
        ],
        "parameters": [
          {
            "name": "width",
            "in": "path",
            "description": "Width of the image",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          },
          {
            "name": "height",
            "in": "path",
            "description": "Height of the image",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Could not find an alpaca matching your requirements"
          }
        }
      }
    }
  }
}